Move filters.py to filter folder
* This is only the Python migration of filter objects
