+++ hoomd/filter/__init__.py
@@ -1,0 +1,5 @@
+from hoomd.filter.filter_ import ParticleFilter
+from hoomd.filter.all_ import All
+from hoomd.filter.set_ import Intersection, SetDifference, Union
+from hoomd.filter.tags import Tags
+from hoomd.filter.type_ import Type
